Adds a test for the cli tool
