+++ .github/workflows/ci.yml
@@ -30,6 +30,8 @@ jobs:
     - uses: actions/checkout@v4
     - run: npm install hargassner_telnet
     - run: sudo node test/TestWithIntegratedSocketServer.js
+    - run: echo "Test command line tool script"
+    - run: ./test/hgboiler-cli-test.bash
 
   security:
     runs-on: ubuntu-latest
